pyproject.toml
[tool.poetry]
name = "datajob-etl"
version = "0.1.0"
description = ""
authors = ["Harrison Zuleta Montoya <harrison.zmontoya@gmail.com>"]
readme = "README.md"
packages = [
    { include = "utils" },
    { include = "db" },
    { include = "actions" },
    { include = "extraction" },
    { include = "transformation" },
    { include = "load" }
]

[tool.poetry.dependencies]
python = "^3.12"
sqlalchemy = "^2.0.48"
pydantic = "^2.12.5"
python-dotenv = "^1.2.2"
pytest = "^9.0.2"
pandas = "^3.0.1"
psycopg2-binary = "^2.9.11"
dbt-core = "^1.11.7"
dbt-postgres = "^1.10.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
